+++ src/icemac/ab/calendar/browser/renderer/tests/test_table.py
@@ -1,5 +1,5 @@
 from datetime import date
-from icemac.ab.calendar.browser.renderer.interfaces import IEventDescription
+from icemac.ab.calendar.browser.interfaces import IEventDescription
 from icemac.ab.calendar.browser.renderer.interfaces import IRenderer
 from icemac.ab.calendar.browser.renderer.interfaces import UnknownLanguageError
 from icemac.ab.calendar.browser.renderer.table import Table, TableEvent